actually undoing all these FullScreenOverlay changes
